+++ sdk/resourcemanager/compute/armcompute/fake/sharedgalleries_server.go
@@ -15,8 +15,7 @@ import (
 	azfake "github.com/Azure/azure-sdk-for-go/sdk/azcore/fake"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/fake/server"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/compute/armcompute/v5"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/compute/armcompute/v6"
 	"net/http"
 	"net/url"
 	"regexp"
@@ -27,27 +26,19 @@ type SharedGalleriesServer struct {
 	// Get is the fake for method SharedGalleriesClient.Get
 	// HTTP status codes to indicate success: http.StatusOK
 	Get func(ctx context.Context, location string, galleryUniqueName string, options *armcompute.SharedGalleriesClientGetOptions) (resp azfake.Responder[armcompute.SharedGalleriesClientGetResponse], errResp azfake.ErrorResponder)
-
-	// NewListPager is the fake for method SharedGalleriesClient.NewListPager
-	// HTTP status codes to indicate success: http.StatusOK
-	NewListPager func(location string, options *armcompute.SharedGalleriesClientListOptions) (resp azfake.PagerResponder[armcompute.SharedGalleriesClientListResponse])
 }
 
 // NewSharedGalleriesServerTransport creates a new instance of SharedGalleriesServerTransport with the provided implementation.
 // The returned SharedGalleriesServerTransport instance is connected to an instance of armcompute.SharedGalleriesClient via the
 // azcore.ClientOptions.Transporter field in the client's constructor parameters.
 func NewSharedGalleriesServerTransport(srv *SharedGalleriesServer) *SharedGalleriesServerTransport {
-	return &SharedGalleriesServerTransport{
-		srv:          srv,
-		newListPager: newTracker[azfake.PagerResponder[armcompute.SharedGalleriesClientListResponse]](),
-	}
+	return &SharedGalleriesServerTransport{srv: srv}
 }
 
 // SharedGalleriesServerTransport connects instances of armcompute.SharedGalleriesClient to instances of SharedGalleriesServer.
 // Don't use this type directly, use NewSharedGalleriesServerTransport instead.
 type SharedGalleriesServerTransport struct {
-	srv          *SharedGalleriesServer
-	newListPager *tracker[azfake.PagerResponder[armcompute.SharedGalleriesClientListResponse]]
+	srv *SharedGalleriesServer
 }
 
 // Do implements the policy.Transporter interface for SharedGalleriesServerTransport.
@@ -64,8 +55,6 @@ func (s *SharedGalleriesServerTransport) Do(req *http.Request) (*http.Response, 
 	switch method {
 	case "SharedGalleriesClient.Get":
 		resp, err = s.dispatchGet(req)
-	case "SharedGalleriesClient.NewListPager":
-		resp, err = s.dispatchNewListPager(req)
 	default:
 		err = fmt.Errorf("unhandled API %s", method)
 	}
@@ -106,55 +95,6 @@ func (s *SharedGalleriesServerTransport) dispatchGet(req *http.Request) (*http.R
 	resp, err := server.MarshalResponseAsJSON(respContent, server.GetResponse(respr).SharedGallery, req)
 	if err != nil {
 		return nil, err
-	}
-	return resp, nil
-}
-
-func (s *SharedGalleriesServerTransport) dispatchNewListPager(req *http.Request) (*http.Response, error) {
-	if s.srv.NewListPager == nil {
-		return nil, &nonRetriableError{errors.New("fake for method NewListPager not implemented")}
-	}
-	newListPager := s.newListPager.get(req)
-	if newListPager == nil {
-		const regexStr = `/subscriptions/(?P<subscriptionId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.Compute/locations/(?P<location>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/sharedGalleries`
-		regex := regexp.MustCompile(regexStr)
-		matches := regex.FindStringSubmatch(req.URL.EscapedPath())
-		if matches == nil || len(matches) < 2 {
-			return nil, fmt.Errorf("failed to parse path %s", req.URL.Path)
-		}
-		qp := req.URL.Query()
-		locationParam, err := url.PathUnescape(matches[regex.SubexpIndex("location")])
-		if err != nil {
-			return nil, err
-		}
-		sharedToUnescaped, err := url.QueryUnescape(qp.Get("sharedTo"))
-		if err != nil {
-			return nil, err
-		}
-		sharedToParam := getOptional(armcompute.SharedToValues(sharedToUnescaped))
-		var options *armcompute.SharedGalleriesClientListOptions
-		if sharedToParam != nil {
-			options = &armcompute.SharedGalleriesClientListOptions{
-				SharedTo: sharedToParam,
-			}
-		}
-		resp := s.srv.NewListPager(locationParam, options)
-		newListPager = &resp
-		s.newListPager.add(req, newListPager)
-		server.PagerResponderInjectNextLinks(newListPager, req, func(page *armcompute.SharedGalleriesClientListResponse, createLink func() string) {
-			page.NextLink = to.Ptr(createLink())
-		})
-	}
-	resp, err := server.PagerResponderNext(newListPager, req)
-	if err != nil {
-		return nil, err
-	}
-	if !contains([]int{http.StatusOK}, resp.StatusCode) {
-		s.newListPager.remove(req)
-		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK", resp.StatusCode)}
-	}
-	if !server.PagerResponderMore(newListPager) {
-		s.newListPager.remove(req)
 	}
 	return resp, nil
 }